requirements.txt
asgiref==3.6.0
Django==4.1.5
Pillow==9.4.0
sqlparse==0.4.3
django-ckeditor==6.5.1